Changed structure of structure
Structures had a duplicated method
Now, this method is on abstract structure
