+++ ForceGamaAttack/src/player/Structure.java
@@ -1,12 +1,14 @@
 package player;
 
+import java.awt.event.KeyEvent;
+
 import jplay.Keyboard;
 import jplay.Sprite;
 
 public abstract class Structure extends Sprite{
 	private int health;
 	private int maxHealth;
-	double speed;
+	protected double speed;
 	int armor;
 	int size;
 	int powerUp;
@@ -47,8 +49,23 @@ public abstract class Structure extends Sprite{
 		this.health -= damage;
 	}
 	
-	public abstract void moveX(double x);
-	public abstract void moveY(double y);
+	public void moveX(double x) {
+		if(this.getKeyboard().keyDown(KeyEvent.VK_RIGHT)) {
+			this.x += x * speed;
+		}
+		else if(this.getKeyboard().keyDown(KeyEvent.VK_LEFT)) {
+			this.x -= x * speed;
+		}
+	}
+	
+	public void moveY(double y) {
+		if(this.getKeyboard().keyDown(KeyEvent.VK_DOWN)) {
+			this.y += y * speed;
+		}
+		else if(this.getKeyboard().keyDown(KeyEvent.VK_UP)) {
+			this.y -= y * speed;
+		}
+	}
 
 	public int getHeight() {
 		return this.height;
