+++ ForceGamaAttack/src/player/StructureStrategyJava.java
@@ -1,7 +1,4 @@
 package player;
-import java.awt.event.KeyEvent;
-
-import jplay.Keyboard;
 
 public class StructureStrategyJava extends Structure {
 	
@@ -14,25 +11,6 @@ public class StructureStrategyJava extends Structure {
 		super(x, y, sprite, javaHealth);
 		this.height = 90;
 		this.width = 46;
-	}
-	
-	@Override
-	public void moveX(double x) {
-		if(this.getKeyboard().keyDown(KeyEvent.VK_RIGHT)) {
-			this.x += x * speed;
-		}
-		else if(this.getKeyboard().keyDown(KeyEvent.VK_LEFT)) {
-			this.x -= x * speed;
-		}
-	}
-	
-	@Override
-	public void moveY(double y) {
-		if(this.getKeyboard().keyDown(KeyEvent.VK_DOWN)) {
-			this.y += y * speed;
-		}
-		else if(this.getKeyboard().keyDown(KeyEvent.VK_UP)) {
-			this.y -= y * speed;
-		}
+		this.speed = 0.5;
 	}
 }
